feat(onboarding): create use-onboarding-visibility hook
